requirements.txt
pytest==8.3.5
playwright==1.50.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
python-dotenv==1.0.1
